students/jongbeom/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.15.0
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.11.0
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
six==1.16.0
sqlparse==0.4.2
